screenshot-version/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
selenium>=4.15.0
Pillow>=10.0.0
webdriver-manager>=4.0.0
google-generativeai>=0.3.0